+++ src/generator/d1/tests/wrangler_tests.rs
@@ -1,11 +1,16 @@
-use common::WranglerFormat;
+use common::{CidlSpec, WranglerFormat};
+use d1::D1Generator;
+
+use anyhow::Result;
+use insta::assert_snapshot;
+
 use std::path::PathBuf;
 
 #[test]
 fn test_serialize_wrangler_spec() {
     // Filled TOML
     {
-        let wrangler_path = PathBuf::from("../fixtures/wrangler.toml");
+        let wrangler_path = PathBuf::from("../../test_fixtures/wrangler.toml");
         WranglerFormat::from_path(&wrangler_path).expect("Wrangler file to serialize");
     }
 
@@ -18,7 +23,7 @@ fn test_serialize_wrangler_spec() {
 
     // Filled JSON
     {
-        let wrangler_path = PathBuf::from("../fixtures/wrangler.json");
+        let wrangler_path = PathBuf::from("../../test_fixtures/wrangler.json");
         WranglerFormat::from_path(&wrangler_path).expect("Wrangler file to serialize");
     }
 
@@ -28,4 +33,29 @@ fn test_serialize_wrangler_spec() {
             .as_spec()
             .expect("Wrangler file to serialize");
     }
+}
+
+#[test]
+fn test_generate_empty_wrangler_snap() -> Result<()> {
+    // Arrange
+    let cidl = {
+        let cidl_path = PathBuf::from("../../test_fixtures/cidl.json");
+        let cidl_contents = std::fs::read_to_string(cidl_path)?;
+        serde_json::from_str::<CidlSpec>(&cidl_contents)?
+    };
+
+    let wrangler = WranglerFormat::Toml(toml::from_str("").unwrap());
+
+    let d1gen = D1Generator::new(cidl, wrangler.as_spec()?);
+
+    // Act
+    let updated_wrangler = d1gen.wrangler();
+
+    // Assert
+    assert_snapshot!(
+        "updated_wrangler_from_empty",
+        format!("{}", toml::to_string_pretty(&updated_wrangler).unwrap())
+    );
+
+    Ok(())
 }
