+++ src/generator/workers/tests/workers_tests.rs
@@ -6,10 +6,10 @@ use workers::WorkersFactory;
 use std::path::PathBuf;
 
 #[test]
-fn test_generate_client_snapshot() -> Result<()> {
+fn test_generate_workers_snapshot() -> Result<()> {
     // Arrange
     let cidl = {
-        let cidl_path = PathBuf::from("../fixtures/cidl.json");
+        let cidl_path = PathBuf::from("../../test_fixtures/cidl.json");
         let cidl_contents = std::fs::read_to_string(cidl_path)?;
         serde_json::from_str::<CidlSpec>(&cidl_contents)?
     };
